Ejercicio 3 Boletin 9

+++ Boletin09/Boletin09_3/Comprobador.java
@@ -2,7 +2,7 @@
 package boletin09_3;
 
 public class Comprobador {
-    public static float comprobarDatos(int num1){
+    public static int comprobarDatos(int num1){
         while(num1<0){
             num1=LerDatos.lerEnteiro("El numero es negativo, introduce otro: ");
         }
